Use brew installed bash as default shell

+++ script/docker.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-curl https://download.docker.com/mac/stable/Docker.dmg -o ~/Downloads/Docker.dmg
-(cd ~/Downloads && open Docker.dmg)
-# Drag and drop icon to Applications folder
+if [[ ! -f '/usr/local/bin/docker' ]]; then
+  curl https://download.docker.com/mac/stable/Docker.dmg -o ~/Downloads/Docker.dmg
+  (cd ~/Downloads && open Docker.dmg)
+  rm -rf ~/Downloads/Docker.dmg
+fi
